Convert tabular network file to gis layer.

The network acquired from google is saved as a shapefile. This function helps converts file from CSV to GIS layer and provides an option to save GIS object as a ".gpkg".

Usage

make_network_link_layer(location, folder, auto_save = F, network_object = NULL)

Arguments

location

character string pointing to top level location where data acquired from Google was saved to.

folder

character string of name where data was automatically saved to from google data download.

auto_save

Boolean (T/F - default F) indicating if you want the GIS layer to be saved. Default just creates an object without saving.

network_object

tabular link data. Default is NULL or input left empty - function will use location and folder inputs to load object and then convert.

Value

GIS layer of network with CRS 4326

Examples

data('replica_queried_network')

head(replica_queried_network)
#>   stableEdgeId startLat  startLon   endLat    endLon          streetName
#> 1 3.314086e+18 47.57149 -122.3263 47.57128 -122.3218 West Seattle Bridge
#> 2 1.336408e+19 47.71454 -122.3264 47.71860 -122.3243                 I 5
#> 3 1.122087e+19 47.67396 -122.3211 47.66535 -122.3220         I 5 Express
#> 4 2.710055e+18 47.65245 -122.3473 47.65227 -122.3473 Aurora Avenue North
#> 5 1.028386e+19 47.56954 -122.3198 47.57020 -122.3199                <NA>
#> 6 6.870997e+18 47.61626 -122.3286 47.61645 -122.3283                <NA>
#>   distance     osmid    speed               flags lanes       highway
#> 1   341000  32289049 51.74856        [ALLOWS_CAR]     2      motorway
#> 2   479000 306194398 88.02476        [ALLOWS_CAR]     6      motorway
#> 3   959000 331932767 88.02476        [ALLOWS_CAR]    -1      motorway
#> 4    20000 457542648 62.11319 [ALLOWS_PEDESTRIAN]     0         trunk
#> 5    73000  32222767 72.47782        [ALLOWS_CAR]     2 motorway_link
#> 6    30000 436165683 72.47782        [ALLOWS_CAR]     1 motorway_link
#>                                                                                                                                                                                                                  geometry
#> 1                                                                      LINESTRING(-122.326328 47.571485, -122.326062 47.571474, -122.324805 47.57126, -122.324484 47.57124, -122.324161 47.571232, -122.321796 47.571279)
#> 2 LINESTRING(-122.326383 47.714536, -122.32561 47.715517, -122.325202 47.716067, -122.324979 47.716434, -122.324666 47.71707, -122.324502 47.717483, -122.324367 47.717924, -122.324282 47.718343, -122.324258 47.718603)
#> 3                                                                     LINESTRING(-122.321138 47.673956, -122.321675 47.672022, -122.321843 47.671213, -122.321914 47.67068, -122.321973 47.669872, -122.322007 47.665353)
#> 4                                                                                                                                                                LINESTRING(-122.347252 47.652447, -122.347254 47.652271)
#> 5                                                                                                                     LINESTRING(-122.319812 47.569542, -122.319788 47.569727, -122.319793 47.569911, -122.31987 47.5702)
#> 6                                                                                                                                         LINESTRING(-122.328613 47.616257, -122.328483 47.616361, -122.328335 47.616447)
#>   flag_contains
#> 1          TRUE
#> 2          TRUE
#> 3          TRUE
#> 4          TRUE
#> 5          TRUE
#> 6          TRUE

make_network_link_layer(
  network_object = replica_queried_network
  ,auto_save = F
)
#> Spatial links were made using supplied tabular network object...
#> Error in st_geometry_type(replica_queried_network): could not find function "st_geometry_type"