Convert tabular first link file to gis layer.

The network acquired from google is saved as a shapefile. This function helps converts file from CSV to GIS layer and provides an option to save GIS object as a ".gpkg".

Usage

make_trip_origin_point_layer(
  location,
  folder,
  auto_save = F,
  first_links_object = NULL
)

Arguments

location

character string pointing to top level location where data acquired from google was saved to.

folder

character string of name where data was automatically saved to from Google data download.

auto_save

boolean (T/F - default F) indicating if you want the GIS layer to be saved. Default just creates an object without saving.

network_object

network object containing links. Default is NULL or input left empty - function will use location and folder inputs to load object and then convert.

Value

GIS layer of network with CRS 4326

Examples

data('replica_queried_network_links')

head(replica_queried_network_links)
#> Simple feature collection with 6 features and 13 fields
#> Geometry type: LINESTRING
#> Dimension:     XY
#> Bounding box:  xmin: -122.3766 ymin: 47.53192 xmax: -122.3003 ymax: 47.68306
#> Geodetic CRS:  WGS 84
#> # A tibble: 6 × 14
#>   stabl…¹ start…² start…³ endLat endLon stree…⁴ dista…⁵  osmid speed flags lanes
#>   <chr>     <dbl>   <dbl>  <dbl>  <dbl> <chr>     <int>  <int> <dbl> <chr> <int>
#> 1 171637…    47.6   -122.   47.6  -122. I 90     224000 6.21e8  62.1 [ALL…     2
#> 2 237485…    47.5   -122.   47.5  -122. 35th A…  160000 3.32e8  36.2 [ALL…    -1
#> 3 600307…    47.7   -122.   47.7  -122. Montla…   32000 5.01e8  46.6 [ALL…     2
#> 4 104579…    47.7   -122.   47.7  -122. East G…   32000 6.41e6  36.2 [ALL…    -1
#> 5 608423…    47.7   -122.   47.7  -122. Green …   32000 8.03e6  36.2 [ALL…     0
#> 6 499973…    47.7   -122.   47.7  -122. Northe…   32000 4.28e8  36.2 [ALL…    -1
#> # … with 3 more variables: highway <chr>, flag_contains <lgl>,
#> #   geom <LINESTRING [°]>, and abbreviated variable names ¹​stableEdgeId,
#> #   ²​startLat, ³​startLon, ⁴​streetName, ⁵​distance

make_network_centroid_layer(
  location = "data/req_dev"
  ,folder = "data_20230106_154119"
  ,network_object = replica_queried_network_links
  ,auto_save = F
)
#> Centroids were made using supplied spatial links network object...
#> Error in sf_object %>% mutate(merge_id = row_number()): could not find function "%>%"