Easily query and download Replica trip and network data using the extent of a map layer.

This function takes a bounding box and google big query inputs and returns data back to you. Function currently only works for data using block groups as it is hardcoded in some of the queries. Changes can be made to the query such that other polygon shapes can be used.

This function creates a log file that records inputs, the query sent to Google, and IDs for each Google table that is created in the multi-step query. All of this information can be accessed at later points in time. Table IDs in log file can be used for auditing purposes or copy and pasted and used in the Google Big Query console GUI.

Usage

query_network_trip_using_bbox(
  bb_network_layer,
  bb_sa_layer,
  query_links = c("highway", "corridor", "road", "motorway", "motorway_link", "trunk",
    "trunk_link", "primary", "primary_link", "secondary", "secondary_link", "tertiary",
    "tertiary_link"),
  customer_name,
  trip_table,
  network_table,
  file_destination,
  query_network = T,
  max_record = 1000
)

Arguments

bb_network_layer

bounding box object detailing extent of road network that will be used to query trips with. This should be made using sf::st_bbox().

bb_sa_layer

bounding box object detailing extent of the study area that will be used to perform aggregations. This should be made using sf::st_bbox(). This input is gernerally larger or identical to [bb_network_layer].

query_links

vector of character strings detailing which type of network links to use in the query. These will be used 1) to depict any network link volume graphics with 2) to filter trips with. For the latter, any trip that uses these road within the study area will be including in any reporting.

customer_name

character string indicating google account that will be billed

trip_table

character string indicating which trip table should be queired - location, year, quarter, day, etc should match network_table. Please include entire table name as well '[customer_name].[trip_table]'

network_table

character string indicating which trip table should be queired - location, year, quarter, day, etc should match trip_table. Please include entire table name as well '[customer_name].[network_table]'

file_destination

character string indicating a directory where you want log files and data to be saved to. A new folder containg function outputs will be made there named using the convention 'data_[sys.datetime]'

query_network

boolean (T/F) indicating if the network should be downloaded at function runtime. Default is 'T'. It can be beneficial to wait and see how many links will be downloaded before netowrk is downloaded, it can also be downloaded latter as log files record big query table IDs.

max_record

integer indicating max number of records will be written out for each data acquired by google - default is 1000, use Inf if you do not want to limit download. Useful if you want to check the data out first before downloading large amounts of data. Google tables are still made in full so they can be manually acquired without rerunning this function - see log files.

Value

does not return any objects in R. Function creates a log file and data to folder destination in CSV format.

Examples

#none