Convert tabular network file to gis layer.

The network acquired from google is saved as a shapefile. This function helps converts file from CSV to GIS layer and provides an option to save GIS object as a ".gpkg".

Usage

make_network_link_layer(location, folder, auto_save = F, network_object = NULL)

Arguments

location

character string pointing to top level location where data acquired from google was saved to.

folder

character string of name where data was automatically saved to from google data download.

auto_save

boolean (T/F - default F) inicating if you want the GIS layer to be saved. Default just creates an object without saving.

network_object

tabular link data. Default is NULL or input left empty - function will use location and folder inputs to load object and then convert.

Value

GIS layer of network with CRS 4326

Examples

data('replica_queried_network')

head(replica_queried_network)
#>   stableEdgeId startLat  startLon   endLat    endLon
#> 1 1.716380e+19 47.59232 -122.3259 47.59303 -122.3231
#> 2 2.374856e+18 47.53192 -122.3766 47.53336 -122.3766
#> 3 6.003071e+18 47.65612 -122.3021 47.65640 -122.3020
#> 4 1.045799e+19 47.68030 -122.3258 47.68002 -122.3258
#> 5 6.084231e+18 47.66153 -122.3467 47.66126 -122.3468
#> 6 4.999734e+18 47.68306 -122.3003 47.68306 -122.3008
#>                     streetName distance     osmid    speed
#> 1                         I 90   224000 621426664 62.11319
#> 2        35th Avenue Southwest   160000 332496923 36.23891
#> 3 Montlake Boulevard Northeast    32000 500504113 46.60353
#> 4  East Green Lake Drive North    32000   6410233 36.23891
#> 5         Green Lake Way North    32000   8026709 36.23891
#> 6        Northeast 75th Street    32000 428091103 36.23891
#>                                          flags lanes   highway
#> 1                                 [ALLOWS_CAR]     2  motorway
#> 2 [ALLOWS_CAR, ALLOWS_BIKE, ALLOWS_PEDESTRIAN]    -1 secondary
#> 3 [ALLOWS_CAR, ALLOWS_BIKE, ALLOWS_PEDESTRIAN]     2   primary
#> 4 [ALLOWS_CAR, ALLOWS_BIKE, ALLOWS_PEDESTRIAN]    -1 secondary
#> 5                          [ALLOWS_PEDESTRIAN]     0 secondary
#> 6 [ALLOWS_CAR, ALLOWS_BIKE, ALLOWS_PEDESTRIAN]    -1 secondary
#>                                                                         geometry
#> 1 LINESTRING(-122.32587 47.592317, -122.324199 47.592718, -122.323077 47.593035)
#> 2                       LINESTRING(-122.376642 47.531919, -122.376609 47.533357)
#> 3                        LINESTRING(-122.302087 47.65612, -122.301999 47.656399)
#> 4 LINESTRING(-122.325841 47.680297, -122.325787 47.680178, -122.325753 47.68002)
#> 5                        LINESTRING(-122.34667 47.661529, -122.346839 47.661261)
#> 6                       LINESTRING(-122.300333 47.683056, -122.300754 47.683057)
#>   flag_contains
#> 1          TRUE
#> 2          TRUE
#> 3          TRUE
#> 4          TRUE
#> 5          TRUE
#> 6          TRUE

make_network_link_layer(
  network_object = replica_queried_network
  ,auto_save = F
)
#> Centroids made using suplied network object...
#> Simple feature collection with 25869 features and 13 fields
#> Geometry type: LINESTRING
#> Dimension:     XY
#> Bounding box:  xmin: -122.4135 ymin: 47.50069 xmax: -122.2406 ymax: 47.74676
#> Geodetic CRS:  WGS 84
#> First 10 features:
#>    stableEdgeId startLat  startLon   endLat    endLon
#> 1  1.716380e+19 47.59232 -122.3259 47.59303 -122.3231
#> 2  2.374856e+18 47.53192 -122.3766 47.53336 -122.3766
#> 3  6.003071e+18 47.65612 -122.3021 47.65640 -122.3020
#> 4  1.045799e+19 47.68030 -122.3258 47.68002 -122.3258
#> 5  6.084231e+18 47.66153 -122.3467 47.66126 -122.3468
#> 6  4.999734e+18 47.68306 -122.3003 47.68306 -122.3008
#> 7  3.453132e+18 47.67190 -122.3174 47.67162 -122.3174
#> 8  1.585985e+19 47.66923 -122.3060 47.66930 -122.3056
#> 9  1.425740e+19 47.66865 -122.3622 47.66840 -122.3620
#> 10 1.718644e+19 47.53488 -122.2806 47.53516 -122.2807
#>                             streetName distance     osmid    speed
#> 1                                 I 90   224000 621426664 62.11319
#> 2                35th Avenue Southwest   160000 332496923 36.23891
#> 3         Montlake Boulevard Northeast    32000 500504113 46.60353
#> 4          East Green Lake Drive North    32000   6410233 36.23891
#> 5                 Green Lake Way North    32000   8026709 36.23891
#> 6                Northeast 75th Street    32000 428091103 36.23891
#> 7              Roosevelt Way Northeast    32000 421652698 36.23891
#> 8          Northeast Ravenna Boulevard    32000 590421615 36.23891
#> 9              Northwest Market Street    32000 158265863 36.23891
#> 10 Martin Luther King Junior Way South    32000 428291925 36.23891
#>                                           flags lanes   highway flag_contains
#> 1                                  [ALLOWS_CAR]     2  motorway          TRUE
#> 2  [ALLOWS_CAR, ALLOWS_BIKE, ALLOWS_PEDESTRIAN]    -1 secondary          TRUE
#> 3  [ALLOWS_CAR, ALLOWS_BIKE, ALLOWS_PEDESTRIAN]     2   primary          TRUE
#> 4  [ALLOWS_CAR, ALLOWS_BIKE, ALLOWS_PEDESTRIAN]    -1 secondary          TRUE
#> 5                           [ALLOWS_PEDESTRIAN]     0 secondary          TRUE
#> 6  [ALLOWS_CAR, ALLOWS_BIKE, ALLOWS_PEDESTRIAN]    -1 secondary          TRUE
#> 7  [ALLOWS_CAR, ALLOWS_BIKE, ALLOWS_PEDESTRIAN]     2 secondary          TRUE
#> 8  [ALLOWS_CAR, ALLOWS_BIKE, ALLOWS_PEDESTRIAN]     1 secondary          TRUE
#> 9                           [ALLOWS_PEDESTRIAN]     0 secondary          TRUE
#> 10 [ALLOWS_CAR, ALLOWS_BIKE, ALLOWS_PEDESTRIAN]     2 secondary          TRUE
#>                          geometry
#> 1  LINESTRING (-122.3259 47.59...
#> 2  LINESTRING (-122.3766 47.53...
#> 3  LINESTRING (-122.3021 47.65...
#> 4  LINESTRING (-122.3258 47.68...
#> 5  LINESTRING (-122.3467 47.66...
#> 6  LINESTRING (-122.3003 47.68...
#> 7  LINESTRING (-122.3174 47.67...
#> 8  LINESTRING (-122.306 47.669...
#> 9  LINESTRING (-122.3622 47.66...
#> 10 LINESTRING (-122.2806 47.53...